Extract files for Flux logic
