+++ src/constants.js
@@ -1,0 +1,2 @@
+export const CLEAR_TRANSCRIPT = 'CLEAR_TRANSCRIPT'
+export const APPEND_TRANSCRIPT = 'APPEND_TRANSCRIPT'
